Refactor Layer and its subcomponents, replacing cloneElement by context
