+++ src/Components/Map/index.tsx
@@ -2,8 +2,6 @@ export { UtopiaMap } from './UtopiaMap'
 export * from './Layer'
 export { Tags } from './Tags'
 export * from './Permissions'
-export { ItemForm } from './ItemForm'
-export { ItemView } from './ItemView'
 export { PopupTextAreaInput } from './Subcomponents/ItemPopupComponents/PopupTextAreaInput'
 export { PopupStartEndInput } from './Subcomponents/ItemPopupComponents/PopupStartEndInput'
 export { PopupTextInput } from './Subcomponents/ItemPopupComponents/PopupTextInput'
