+++ src/Components/Profile/index.tsx
@@ -2,3 +2,5 @@ export { UserSettings } from './UserSettings'
 // export { PlusButton } from './Subcomponents/PlusButton'
 export { ProfileView } from './ProfileView'
 export { ProfileForm } from './ProfileForm'
+export { CardForm } from './templateComponents/CardForm'
+export { CardView } from './templateComponents/CardView'
