+++ src/Components/Profile/templateComponents/TemplateItemContext.ts
@@ -1,0 +1,7 @@
+import { createContext } from 'react'
+
+import type { Item } from '#types/Item'
+
+const ItemContext = createContext<Item | undefined>(undefined)
+
+export default ItemContext
